fix bag start

+++ src/main/java/com/example/amagazishi/dto/ProductRequestRegister.java
@@ -13,6 +13,6 @@ public class ProductRequestRegister {
     private Long price;
     private String title;
     private String compound;
-    private Integer discount;
+    private Long discount;
     private String description;
 }
